Made Signup Working
