+++ backend/models/student.model.js
@@ -2,11 +2,11 @@ import mongoose from 'mongoose';
 
 const studentSchema = new mongoose.Schema(
   {
-    name: {
-      type: String,
-      required: true,
-      trim: true,
-    },
+    // name: {
+    //   type: String,
+    //   required: true,
+    //   trim: true,
+    // },
     rollno: {
       type: String,
       required: true,
